trans removeSearch.js to removeArticle.ts
